LPD-72607 Extract
